Rename `kubernetes` processor -> `add_kubernetes_metadata`
