+++ filebeat/processor/add_kubernetes_metadata/indexing_test.go
@@ -1,4 +1,4 @@
-package kubernetes
+package add_kubernetes_metadata
 
 import (
 	"fmt"
