+++ libbeat/processors/add_kubernetes_metadata/_meta/fields.yml
