+++ metricbeat/processor/add_kubernetes_metadata/indexing_test.go
@@ -1,15 +1,16 @@
-package kubernetes
+package add_kubernetes_metadata
 
 import (
 	"fmt"
 	"testing"
 
-	"github.com/elastic/beats/libbeat/common"
-	"github.com/elastic/beats/libbeat/processors/kubernetes"
 	"github.com/stretchr/testify/assert"
+
+	"github.com/elastic/beats/libbeat/common"
+	"github.com/elastic/beats/libbeat/processors/add_kubernetes_metadata"
 )
 
-var metagen = &kubernetes.GenDefaultMeta{}
+var metagen = &add_kubernetes_metadata.GenDefaultMeta{}
 
 func TestIpPortIndexer(t *testing.T) {
 	var testConfig = common.NewConfig()
@@ -22,19 +23,19 @@ func TestIpPortIndexer(t *testing.T) {
 	container := "container"
 	ip := "1.2.3.4"
 	port := int64(80)
-	pod := kubernetes.Pod{
-		Metadata: kubernetes.ObjectMeta{
+	pod := add_kubernetes_metadata.Pod{
+		Metadata: add_kubernetes_metadata.ObjectMeta{
 			Name:      podName,
 			Namespace: ns,
 			Labels: map[string]string{
 				"labelkey": "labelvalue",
 			},
 		},
-		Spec: kubernetes.PodSpec{
-			Containers: make([]kubernetes.Container, 0),
+		Spec: add_kubernetes_metadata.PodSpec{
+			Containers: make([]add_kubernetes_metadata.Container, 0),
 		},
 
-		Status: kubernetes.PodStatus{
+		Status: add_kubernetes_metadata.PodStatus{
 			PodIP: ip,
 		},
 	}
@@ -53,10 +54,10 @@ func TestIpPortIndexer(t *testing.T) {
 		},
 	}
 
-	pod.Spec.Containers = []kubernetes.Container{
+	pod.Spec.Containers = []add_kubernetes_metadata.Container{
 		{
 			Name: container,
-			Ports: []kubernetes.ContainerPort{
+			Ports: []add_kubernetes_metadata.ContainerPort{
 				{
 					Name:          container,
 					ContainerPort: port,
